testing multiple moives components
Versions 3 and 4 covered the creation of the second component of the program - getting the rating system to work for multiple movies. In this commit, I tested to make sure that this component was working before I moved on - that the buttons were working as expected, and that ratings saved for each movie.